+++ src/gr/sampling.h
@@ -127,7 +127,7 @@ private:
             
             for(const auto& p : inputset) {
                 uint64_t& ind = hash[PointType(p)];
-                if (ind >= num_input) {
+                if (ind >= uint64_t(num_input)) {
                 output.push_back( p );
                 ind = output.size();
                 }
